backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
marshmallow==3.22.0
packaging==24.1
pillow==10.4.0
PyJWT==2.9.0
reportlab==4.2.2
requests==2.32.3
sqlparse==0.5.1
sslcommerz-lib==1.0
sslcommerz-sdk==1.0.4
stripe==10.12.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
